requirements.txt
absl-py==1.3.0
attrs==22.1.0
bidict==0.22.0
click==8.1.3
contourpy==1.0.6
cycler==0.11.0
Flask==2.2.2
Flask-Cors==3.0.10
Flask-SocketIO==5.3.2
flatbuffers==22.11.23
fonttools==4.38.0
itsdangerous==2.1.2
Jinja2==3.1.2
kiwisolver==1.4.4
MarkupSafe==2.1.1
matplotlib==3.6.2
mediapipe==0.9.0
numpy==1.23.5
opencv-contrib-python==4.6.0.66
opencv-python==4.6.0.66
packaging==21.3
Pillow==9.3.0
protobuf==3.20.3
pyparsing==3.0.9
python-dateutil==2.8.2
python-engineio==4.3.4
python-socketio==5.7.2
six==1.16.0
Werkzeug==2.2.2
